refactor: move graphql types to separate files
