+++ graphql/schema/user.js
@@ -1,0 +1,21 @@
+const { userData } = require('./util');
+
+module.exports = {
+  User: `
+    type User {
+        _id: ID!
+        password: String!
+        ${userData}
+    }
+    `,
+  AuthUser: `
+    type AuthData {
+        userId: ID!
+        ${userData}
+    }`,
+  UserInputData: `
+    input UserInputData {
+        password: String!
+        ${userData}
+    }`,
+};
