+++ graphql/schema/util.js
@@ -1,0 +1,7 @@
+module.exports = {
+  userData: `
+    email: String!
+    name: String!
+    lastname: String!
+    phonenumber: String!`,
+};
